Ignore events sent to a window that isn't in the map.
First of all, these cannot occur before the window is added to the
map, because both event polling and window creation happen in the main
thread, so there's no race condition. Second, when deleting a window,
if there are residual events, then these should be ignored. Third, it
simplifies the code.